update to use api v4
instead of implementing our own api wrapper, we'll just build on
top of the official api.

the new v4 api is especially nice because it has dropped the janky
xml stuff and produces json by default.

unfortunately, the whole thing is still very API-like and not very
OO like, so this gem would probably be geared towards making the api
more object oriented in nature (with sane defaults)